Don't send position data if both lat and long is 0
Flight Simulator 2020 sends position data if you ask for it, even if you're just poking around in the main menu. When you're not flying or at an airport the coordinates being sent are 00N000W. By filtering these packets out, the chances of getting garbage in your flight log are diminished.